+++ test/Step.test.ts
@@ -6,14 +6,19 @@ describe('simple Cypher', () => {
   afterEach(() => {
     cypher.cleanUp()
   })
+  it('Produce: [MATCH (n) RETURN n]', () => {
+    const actual = cypher.match().return().getCypher()
+
+    expect(actual).toBe('MATCH (n) RETURN n')
+  })
   it('Produce: [MATCH (n:Person) RETURN n]', () => {
-    const actual = cypher.match([database.Labels.Person]).return().getCypher()
+    const actual = cypher.match(database.Labels.Person).return().getCypher()
 
     expect(actual).toBe('MATCH (n:Person) RETURN n')
   })
   it('Produce: [MATCH (n:Person:Animal) RETURN n]', () => {
     const labels = [database.Labels.Person, database.Labels.Animal]
-    const actual = cypher.match(labels).return().getCypher()
+    const actual = cypher.match(...labels).return().getCypher()
 
     expect(actual).toBe('MATCH (n:Person:Animal) RETURN n')
   })