+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - documentation
 - make Label a class
+### Breaking Change
+- match step takes spread array instead of array
 
 ## [0.0.1]  2022-07-17
 ### Added